pyproject.toml
[project]
name = "task-tracker-apis"
version = "0.1.0"
description = ""
authors = [
    {name = "Harshit ",email = "pawarharshit358@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "ruff (>=0.13.2,<0.14.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "ipython (>=9.5.0,<10.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "uuid (>=1.30,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
